Quick tests for mouse input handling on the game side

+++ engine/src/main/java/fi/jakojaannos/roguelite/engine/input/AxialInput.java
@@ -4,9 +4,9 @@ import lombok.Getter;
 
 public class AxialInput {
     @Getter private final InputAxis axis;
-    @Getter private final float value;
+    @Getter private final double value;
 
-    public AxialInput(InputAxis axis, float value) {
+    public AxialInput(InputAxis axis, double value) {
         this.axis = axis;
         this.value = value;
     }
